fix: zero clippy warnings, add docs, bump to v0.1.1, fix gitignore
- Fix all 47 clippy warnings in edgeparse-core (dead code, loop vars,
  match->if-let, Vec->slice, clamp, complex type, missing docs)
- Add doc comments to 25 struct/enum fields across 5 source files
- Create hybrid.rs placeholder to fix cargo fmt compatibility
- Update Makefile lint/fmt targets to scope to publishable crates
- Bump workspace version 0.1.0 -> 0.1.1 (fixes crates.io repo URL)
- Fix .gitignore: separate logs/ and .playwright-mcp/ entries
- Apply cargo fmt to edgeparse-core and edgeparse-cli

+++ crates/edgeparse-core/src/pdf/annotation_extractor.rs
@@ -147,7 +147,7 @@ fn get_rect(dict: &lopdf::Dictionary) -> Option<[f64; 4]> {
     let mut result = [0.0f64; 4];
     for (i, obj) in arr.iter().enumerate().take(4) {
         result[i] = match obj {
-            Object::Real(f) => f64::from(*f),
+            Object::Real(f) => *f,
             Object::Integer(i) => *i as f64,
             _ => return None,
         };
